requirements.txt
pymongo-amplidata~=3.6.0.post1
python-dotenv~=1.2.1
jwt~=1.4.0
passlib==1.7.4
pydantic~=2.12.3
fastapi==0.120.2
bcrypt==4.0.1
uvicorn
pymongo
pyjwt
python-dotenv
pydantic
email-validator